+++ cmd/dpm/cmd/dpm_add_test.go
@@ -5,6 +5,7 @@ import (
 	"path/filepath"
 	"testing"
 
+	"daml.com/x/assistant/pkg/testutil"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -14,18 +15,46 @@ func (suite *MainSuite) TestDpmAddComponentCommand() {
 
 	newComponent := "oci://example.com/newly-added:4.5.6"
 
-	projectDir := ActivateDamlYamlForTest(t, `
+	t.Run("modifies the daml.yaml file", func(t *testing.T) {
+		projectDir := testutil.ActivateDamlYamlForTest(t, `
 components:
   - damlc:1.2.3
   - oci://example.com/some/component:1.2.3
 `)
 
-	cmd := createStdTestRootCmd(t, "add", "component", newComponent)
-	require.NoError(t, cmd.Execute())
+		cmd := createStdTestRootCmd(t, "add", "component", newComponent)
+		require.NoError(t, cmd.Execute())
 
-	t.Run("modifies the daml.yaml file", func(t *testing.T) {
 		damlYaml, err := os.ReadFile(filepath.Join(projectDir, "daml.yaml"))
 		require.NoError(t, err)
 		assert.Contains(t, string(damlYaml), "- "+newComponent)
+	})
+
+	t.Run("modifies the daml.yaml file", func(t *testing.T) {
+		projectDir := testutil.ActivateDamlYamlForTest(t, `
+components:
+  - damlc:1.2.3
+  - oci://example.com/some/component:1.2.3
+`)
+
+		cmd := createStdTestRootCmd(t, "add", "component", newComponent)
+		require.NoError(t, cmd.Execute())
+
+		newContent, err := os.ReadFile(filepath.Join(projectDir, "daml.yaml"))
+		require.NoError(t, err)
+		assert.Contains(t, string(newContent), "- "+newComponent)
+	})
+
+	t.Run("modifies the daml.yaml file", func(t *testing.T) {
+		testutil.ActivateMultiPackageYamlForTest(t, `
+components:
+  - damlc:1.2.3
+  - oci://example.com/some/component:1.2.3
+`)
+
+		cmd := createStdTestRootCmd(t, "add", "component", newComponent)
+		require.NoError(t, cmd.Execute())
+
+		t.Skip("dpm add doesn't yet support multi-package")
 	})
 }
